Be stricter when parsing angle strings
A valid angle string now has the format of [number[:number[:number]]], which is 0 to 3 valid number strings separated by colons. If any number string is empty, it is replaced by the value 0. If any number string is non-empty but invalid, a ValueError is raised instead. This replaces the previous relaxed behaviour where the string parsing always succeeds and invalid strings silently result in angles of 0.