Make the str() of ephem.Date round, not truncate
A single test was failing on Travis, and only under Python 2.7, but when
I looked it revealed that dates were not being rounded to the nearest
second but instead truncated to the nearest second when printed out.
This meant that if you put in a date to a number of seconds, it might be
printed as one second earlier, if the rounding that takes place during
input happened to push it back ever so slightly into the previous second
(14.9999 instead of 15.0, for example).  So, taking lessons learned when
building Skyfield and making it agree perfectly with more standard
sources of astronomy data, I am now bumping the date value up by a
half-second before doing the truncating math that produces the string.