CMake config for general-use shared library (#129)
This PR adds a CMake configuration whose aim is to build a shared
libastro C library. This library can be used for example in C or C++
applications. It was tested to work on Linux, Windows, and MacOS X. All
functions are exported. The Python build process is not altered.