Travis CI: remove debugging command, and tox
The tox command, years later, is still broken for Python 2.7: if the
top-level directory of a project contains an importable copy of your
package, then it will be used in preference to the package that tox has
painstakingly built into its virtualenv.  In this case the symptom is
that “_libastro.so” is missing and cannot be imported.